add can delete record

+++ app/src/main/java/de/vogella/android/myapplication/mainpage_Fragment/record_page/Item_Adapter.kt
@@ -1,6 +1,7 @@
 package de.vogella.android.myapplication.mainpage_Fragment.record_page.item_Fragment
 
 import android.os.Bundle
+import android.util.Log
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
@@ -9,11 +10,17 @@ import android.widget.TextView
 import androidx.fragment.app.Fragment
 import androidx.fragment.app.FragmentManager
 import androidx.recyclerview.widget.RecyclerView
+import com.android.volley.Response
 import de.vogella.android.myapplication.R
+import de.vogella.android.myapplication.components.requestQueue_Manager
 import de.vogella.android.myapplication.mainpage_Fragment.insert_page.insert_expend_Fragment
 import de.vogella.android.myapplication.mainpage_Fragment.insert_page.insert_income_Fragment
+import de.vogella.android.myapplication.record_itemInformationActivity
+import org.json.JSONObject
 
-class item_Adapter (private val fa: FragmentManager,private val is_income:Boolean, private val dataSet: ArrayList<ArrayList<String>>):
+class item_Adapter (private val fam: FragmentManager,private val fa: Fragment,
+                    private val is_income:Boolean,
+                    private val dataSet: ArrayList<ArrayList<String>>):
     RecyclerView.Adapter<item_Adapter.ViewHolder>(){
     class ViewHolder(view: View) : RecyclerView.ViewHolder(view) {
         val edit_imageView = view.findViewById(R.id.item_edit) as ImageView
@@ -49,7 +56,7 @@ class item_Adapter (private val fa: FragmentManager,private val is_income:Boolea
         }
 
         viewHolder.edit_imageView.setOnClickListener{
-                val transaction = fa.beginTransaction()
+                val transaction = fam.beginTransaction()
                 transaction.addToBackStack("firstPage")
                 val bundle = Bundle()
                 bundle.putStringArrayList("data",dataSet[position])
@@ -64,8 +71,15 @@ class item_Adapter (private val fa: FragmentManager,private val is_income:Boolea
                 transaction.commit()
         }
 
-        viewHolder.delete_imageView.setOnClickListener{
-            deleteItem(position)
+        viewHolder.delete_imageView.setOnClickListener{ v->
+            val url = "http://10.0.2.2:3001/trade?tid="+dataSet[position][5]
+            val requestManage = requestQueue_Manager(v.context)
+            requestManage.Request("delete",url,null,Response.Listener<JSONObject>{ res ->
+                if (res.getBoolean("is_delete")) {
+                    deleteItem(position)
+                    //fa.activity?.finish()
+                }
+            })
         }
     }
